refactor(CodeEditor): better names
