+++ src/app/types/CodeEditor.ts
@@ -1,3 +1,5 @@
+import { TokenLocation } from "../../ERDoc/types/parser/TokenLocation";
+
 // from 'monaco-editor', importing it directly causes a conflict with SSR
 export enum MarkerSeverity {
   Hint = 1,
@@ -5,3 +7,8 @@ export enum MarkerSeverity {
   Warning = 4,
   Error = 8,
 }
+
+export type ErrorMessage = {
+  errorMessage: string;
+  location: TokenLocation;
+};
