--- src/app/types/ErrorMessage.ts
@@ -1,6 +1,0 @@
-import { TokenLocation } from "../../ERDoc/types/parser/TokenLocation";
-
-export type ErrorMessage = {
-  errorMessage: string;
-  location: TokenLocation;
-};
